🎨 Move reusable parts off test_server to conftest
